@@ -1,2 +1,1 @@
-__all__ = ["__version__"]
-__version__ = "0.1.0"
+__all__ = ["data", "models", "losses", "training", "slam", "geometry", "utils"]
