@@ -1,19 +1,17 @@
 from __future__ import annotations
-from typing import Optional, List, Tuple
-
+from typing import Optional, Tuple, List
 import numpy as np
 import torch
 import cv2
 
 from dino_slam3.models.network import LocalFeatureNet
-from dino_slam3.slam.keypoints import extract_keypoints_and_descriptors, to_cv2
-
+from dino_slam3.slam.keypoints_torch import extract_keypoints_torch
 
 class DinoSLAM3FeatureExtractor:
     """
-    OpenCV-like feature extractor with correct coordinate handling.
+    OpenCV-like feature extractor: detectAndCompute(image, mask=None) -> (keypoints, descriptors)
+    Returns float descriptors (L2) suitable for BFMatcher(NORM_L2).
     """
-
     def __init__(
         self,
         model: LocalFeatureNet,
@@ -39,6 +37,10 @@ class DinoSLAM3FeatureExtractor:
             ckpt = torch.load(checkpoint_path, map_location="cpu")
             self.model.load_state_dict(ckpt["model"], strict=False)
 
+        if self.device.type == "cuda":
+            torch.backends.cuda.matmul.allow_tf32 = True
+            torch.backends.cudnn.allow_tf32 = True
+
     def _pad_to_multiple(self, img: np.ndarray) -> Tuple[np.ndarray, int, int]:
         H, W = img.shape[:2]
         pad_bottom = (self.pad_to - (H % self.pad_to)) % self.pad_to
@@ -50,7 +52,6 @@ class DinoSLAM3FeatureExtractor:
 
     @torch.no_grad()
     def detectAndCompute(self, image: np.ndarray, mask=None):
-        # Accept grayscale or BGR uint8
         if image.ndim == 2:
             image = np.stack([image] * 3, axis=-1)
         if image.shape[2] != 3:
@@ -58,42 +59,33 @@ class DinoSLAM3FeatureExtractor:
 
         H0, W0 = image.shape[:2]
         img_rgb = image[..., ::-1].copy()  # BGR->RGB
-
         img_rgb, pad_bottom, pad_right = self._pad_to_multiple(img_rgb)
-        Hp, Wp = img_rgb.shape[:2]
 
-        img_t = torch.from_numpy(img_rgb).float() / 255.0
-        img_t = img_t.permute(2, 0, 1).unsqueeze(0)  # 1,3,H,W
+        img_t = torch.from_numpy(img_rgb).float().permute(2, 0, 1).unsqueeze(0) / 255.0
 
+        # DINO normalization (ImageNet mean/std)
         mean = torch.tensor([0.485, 0.456, 0.406]).view(1, 3, 1, 1)
         std = torch.tensor([0.229, 0.224, 0.225]).view(1, 3, 1, 1)
         img_t = (img_t - mean) / std
         img_t = img_t.to(self.device)
 
         out = self.model(img_t)
-        kp_batch = extract_keypoints_and_descriptors(
+        k = extract_keypoints_torch(
             out.heatmap, out.desc, out.offset, out.reliability,
             stride=self.stride,
             nms_radius=self.nms_radius,
-            max_keypoints=self.max_keypoints,
             tile_size=self.tile_size,
             k_per_tile=self.k_per_tile,
-        )[0]
+            max_keypoints=self.max_keypoints,
+            valid_mask_img=None,
+        )
+        xy = k.xy_img[0].detach().cpu().numpy().astype(np.float32)
+        desc = k.desc[0].detach().cpu().numpy().astype(np.float32)
+        score = k.score[0].detach().cpu().numpy().astype(np.float32)
 
-        # Map keypoints from padded coords back to original
-        kpts = kp_batch.kpts.copy()
-        kpts[:, 0] = np.clip(kpts[:, 0], 0, Wp - 1)
-        kpts[:, 1] = np.clip(kpts[:, 1], 0, Hp - 1)
-        # Since we padded only bottom/right, original top-left aligns.
-        # Just drop points that fall in padded area.
-        keep = (kpts[:, 0] < W0) & (kpts[:, 1] < H0)
-        kpts = kpts[keep]
-        desc = kp_batch.desc[keep]
-        scores = kp_batch.scores[keep]
+        # drop padded area
+        keep = (xy[:, 0] < W0) & (xy[:, 1] < H0) & (score > 0)
+        xy, desc, score = xy[keep], desc[keep], score[keep]
 
-        kp_batch.kpts = kpts
-        kp_batch.desc = desc
-        kp_batch.scores = scores
-
-        kps, desc_cv = to_cv2(kp_batch)
-        return kps, desc_cv
+        kps = [cv2.KeyPoint(float(x), float(y), 1.0, _response=float(s)) for (x, y), s in zip(xy, score)]
+        return kps, desc
