@@ -1,8 +1,6 @@
 from __future__ import annotations
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict
-import copy
 import yaml
 
 def _deep_update(dst: Dict[str, Any], src: Dict[str, Any]) -> Dict[str, Any]:
@@ -16,11 +14,10 @@ def _deep_update(dst: Dict[str, Any], src: Dict[str, Any]) -> Dict[str, Any]:
 def load_config(path: str | Path) -> Dict[str, Any]:
     path = Path(path)
     cfg = yaml.safe_load(path.read_text())
-    if isinstance(cfg, dict) and "base" in cfg and cfg["base"]:
+    if isinstance(cfg, dict) and cfg.get("base"):
         base_path = (path.parent / cfg["base"]).resolve()
         base_cfg = load_config(base_path)
-        merged = _deep_update(base_cfg, {k: v for k, v in cfg.items() if k != "base"})
-        return merged
+        return _deep_update(base_cfg, {k: v for k, v in cfg.items() if k != "base"})
     return cfg
 
 def ensure_dir(p: str | Path) -> Path:
